+++ scripts/macros/classFeatures/artificer/infusions/radiantWeapon.js
@@ -3,7 +3,11 @@ import {queue} from '../../../../utility/queue.js';
 async function item({speaker, actor, token, character, item, args, scope, workflow}) {
     let infusionId = workflow.item.flags['chris-premades']?.feature?.infusion?.radiantWeapon?.id;
     if (!infusionId) {
-        let validWeapons = workflow.actor.items.filter(i => i.type === 'weapon' && !i.system.properties?.mgc && (i.system.weaponType === 'martialM' || i.system.weaponType === 'martialS'));
+        let validWeapons = workflow.actor.items.filter(i => i.type === 'weapon' && !i.system.properties?.mgc);
+        if (chris.getItem(actor, 'Armor Modifications')) {
+            let gauntlets = chris.getItem(workflow.actor, 'Guardian Armor: Thunder Gauntlets');
+            if (gauntlets) validWeapons.push(gauntlets);
+        }
         if (validWeapons.length === 0) {
             ui.notifications.info('No valid weapon to infuse!');
             return;